Membuat komponen navbar, footer dan memperbaiki tampilan events
