+++ src/app/(auth)/register/page.js
@@ -2,5 +2,4 @@ import { Register } from "@/components/auth/register";
 
 export default function Page() {
   return <Register />;
-  n;
 }
