+++ src/app/(platform)/events/page.js
@@ -1,5 +1,13 @@
 import EventsPage from "@/components/auth/events";
+import Navbar from "@/components/auth/navbar";
+import Footer from "@/components/auth/footer";
 
 export default function Page() {
-  return <EventsPage />;
+  return (
+    <div>
+      <Navbar />
+      <EventsPage />
+      <Footer />
+    </div>
+  );
 }
