+++ src/components/auth/footer.jsx
@@ -1,0 +1,11 @@
+import React from "react";
+
+const Footer = () => {
+  return (
+    <footer className="py-8 px-8 bg-blue-500 text-white text-center">
+      &copy; 2024 oleh Tim Devdynamos @devscale.id
+    </footer>
+  );
+};
+
+export default Footer;
